finpol/backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain & AI
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16

# FAISS
faiss-cpu==1.7.4
numpy==1.26.3

# Environment
python-dotenv==1.0.0

# Utils
python-dateutil==2.8.2
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
